requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
constantly==23.10.4
contextlib2==21.6.0
cryptography==45.0.4
daphne==4.2.0
Django==4.2.23
django-channels==0.7.0
django-cors-headers==4.7.0
django-mongoengine==0.5.6
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
injector==0.22.0
mongoengine==0.29.1
oauthlib==3.2.2
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pymongo==4.13.1
pyOpenSSL==25.1.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
zope.interface==7.2
